Backend/requirements.txt
# Core Motia dependency
motia>=0.5.12

# Common dependencies
requests>=2.28.0
pydantic>=1.10.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# LearnLive AI Feature dependencies
google-generative-ai
python-dotenv
sentence-transformers
faiss-cpu
nltk
